Start using output for normalization gradient
Saving the output of the normalization instead of the input reduces the
memory cost in modern networks, where the output is going to be saved
anyways (e.g., a Linear layer) and the input is only needed here
otherwise.

We do this by changing the norm backward kernels to load from output
instead of the input, and to compute normalized tensor from output
instead of input. To stabilize gradients, we also clamp by magnitude
the gamma value for division.

However, for now it seems that it comes at a price at a somewhat lower
numerical precision. To be investiagated further.

For now we pass the operator tests if we increase the tolerance for
layer normalization's dgamma and dx.

Signed-off-by: Rui Wang <rui@helixon.com>