Make Norms compute backward with output
Saving the output of the normalization instead of the input reduces the
memory cost in modern networks, where the output is going to be saved
anyways (e.g., a Linear layer) and the input is only needed here
otherwise.

We do this by changing the norm backward kernels to load from output
instead of the input, and to compute normalized tensor from output
instead of input.  To stabilize gradients, we also clamp by magnitude
the gamma value for division.

For now we passed LayerNorm and RMSNorm tests with higher numerical
differences which is expected as we use output to compute the gradient
and the numerical errors propagate that way.

Signed-off-by: Rui Wang <rui@helixon.com>